Refactor, cleanup, and release preparation for v0.1.0
Bug fixes:
- Fix SimpleDescRsp::parse off-by-one in cluster list bounds check
- Clamp transition_time to prevent overflow from negative/large values
- Replace unimplemented!() EZSP panic with proper Error::Config
- Handle FrameType::from_cmd0 values 4-7 (default to Poll)
- Fix illuminance comparison: v > 0.0 instead of v > 0 as f64

Code quality:
- Deduplicate parse_ieee_addr into IeeeAddr::from_hex() method
- Static cluster handler dispatch (no Box allocation per message)
- Use IeeeAddr type in CoordinatorEvent (not raw [u8; 8])
- Move bridge.rs into library crate for testability
- Remove Direction enum (single variant, unused parameter)
- Add #[must_use] to key query methods
- Replace all silent .ok() with warn!() logging on errors
- Clean up unused mqtt cmd_tx_clone variable
- Migrate from deprecated serde_yaml to serde_yml

Documentation:
- README.md with build instructions, config examples, MQTT topics,
  migration guide, device support table, architecture overview
- AGENTS.md for AI assistants with critical constraints, common
  pitfalls, architecture decisions, and key files

137 tests, 0 warnings, 0 C dependencies.

+++ src/zigbee/zcl/clusters/illuminance.rs
@@ -17,7 +17,7 @@ impl ClusterHandler for IlluminanceCluster {
         for r in reports {
             if r.attr_id == MEASURED_VALUE {
                 if let Some(v) = r.value.as_f64() {
-                    if v > 0 as f64 && v < 0xFFFF as f64 {
+                    if v > 0.0 && v < 0xFFFF as f64 {
                         let lux = f64::powf(10.0, (v - 1.0) / 10_000.0);
                         out.push(("illuminance".into(), json!(lux.round() as u32)));
                         out.push(("illuminance_lux".into(), json!(lux.round() as u32)));
